Add a preferred time unit config setting.
This commit adds a preferred time unit setting to the surfer config.
This is useful when you have waveforms that are generated from a
simulator based on the precision of the simulation but you generally
prefer to view waveforms at a different time unit.

The default value for the new setting is "Auto" to preserve current
behavior when the setting is not present in the user's config or there
is no user config.